chore: upate tests
