+++ packages/legacy/core/__mocks__/custom/@react-navigation/native.ts
@@ -1,5 +1,27 @@
-const useIsFocused = jest.fn()
 const useFocusEffect = jest.fn()
 const createNavigatorFactory = jest.fn()
+const navigate = jest.fn()
 
-export { useIsFocused, useFocusEffect, createNavigatorFactory }
+const navigation = {
+  navigate,
+  setOptions: jest.fn(),
+  getParent: jest.fn(() => ({
+    navigate,
+  })),
+  getState: jest.fn(() => ({
+    index: jest.fn(),
+  })),
+  goBack: jest.fn(),
+  pop: jest.fn(),
+  reset: jest.fn(),
+}
+
+const useNavigation = () => {
+  return navigation
+}
+
+const useIsFocused = () => {
+  return true
+}
+
+export { navigate, useNavigation, useIsFocused, useFocusEffect, createNavigatorFactory }
