+++ packages/legacy/core/__tests__/components/CameraDisclosureModal.test.tsx
@@ -1,4 +1,4 @@
-import { useNavigation } from '@react-navigation/core'
+import { useNavigation } from '@react-navigation/native'
 import { render, fireEvent, act } from '@testing-library/react-native'
 import React from 'react'
 
@@ -7,9 +7,7 @@ import { Screens } from '../../App/types/navigators'
 import { testIdWithKey } from '../../App/utils/testable'
 
 let requestCameraUse = jest.fn(() => Promise.resolve(true))
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
+
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
