+++ packages/legacy/core/__tests__/components/HeaderRightHome.test.tsx
@@ -3,9 +3,6 @@ import React from 'react'
 
 import HeaderRightHome from '../../App/components/buttons/HeaderHome'
 
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
