+++ packages/legacy/core/__tests__/screens/ListContacts.test.tsx
@@ -1,4 +1,4 @@
-import { useNavigation } from '@react-navigation/core'
+import { useNavigation } from '@react-navigation/native'
 import { act, fireEvent, render } from '@testing-library/react-native'
 import React from 'react'
 import ListContacts from '../../App/screens/ListContacts'
@@ -7,9 +7,6 @@ import { ConnectionRecord, DidExchangeRole, DidExchangeState } from '@aries-fram
 import { ConfigurationContext } from '../../App/contexts/configuration'
 import configurationContext from '../contexts/configuration'
 
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
