+++ packages/legacy/core/__tests__/screens/Onboarding.test.tsx
@@ -9,9 +9,6 @@ import { OnboardingTheme, theme } from '../../App/theme'
 
 export const carousel: OnboardingStyleSheet = createCarouselStyle(OnboardingTheme)
 
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
