+++ packages/legacy/core/__tests__/screens/ProofDetails.test.tsx
@@ -7,7 +7,7 @@ import {
 import { Attachment, AttachmentData } from '@aries-framework/core/build/decorators/attachment/Attachment'
 import { useProofById } from '@aries-framework/react-hooks'
 import mockRNCNetInfo from '@react-native-community/netinfo/jest/netinfo-mock'
-import { useNavigation } from '@react-navigation/core'
+import { useNavigation } from '@react-navigation/native'
 import '@testing-library/jest-native/extend-expect'
 import { act, cleanup, fireEvent, render, RenderAPI } from '@testing-library/react-native'
 import React from 'react'
@@ -22,9 +22,6 @@ import { ConfigurationContext } from '../../App/contexts/configuration'
 
 jest.mock('react-native/Libraries/EventEmitter/NativeEventEmitter')
 jest.mock('@react-native-community/netinfo', () => mockRNCNetInfo)
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
