+++ packages/legacy/core/__tests__/screens/Scan.test.tsx
@@ -1,13 +1,10 @@
-import { useNavigation } from '@react-navigation/core'
+import { useNavigation } from '@react-navigation/native'
 import { render, waitFor } from '@testing-library/react-native'
 import React from 'react'
 
 import Scan from '../../App/screens/Scan'
 
 jest.mock('react-native-permissions', () => require('react-native-permissions/mock'))
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
