+++ packages/legacy/core/__tests__/screens/UseBiometry.test.tsx
@@ -7,9 +7,6 @@ import { testIdWithKey } from '../../App/utils/testable'
 import authContext from '../contexts/auth'
 import timeTravel from '../helpers/timetravel'
 
-jest.mock('@react-navigation/core', () => {
-  return require('../../__mocks__/custom/@react-navigation/core')
-})
 jest.mock('@react-navigation/native', () => {
   return require('../../__mocks__/custom/@react-navigation/native')
 })
